% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{ci_eta2}
\alias{ci_eta2}
\title{Confidence interval for eta-squared / via noncentral chi-square inversion}
\usage{
ci_eta2(eta2, u, D0, conf.level = 0.95)
}
\arguments{
\item{eta2}{numeric. the effect size index for predictor x}

\item{u}{integer. Degrees of freedom associated with predictor x.}

\item{D0}{numeric. Null-model deviance. Required for eta2;}

\item{conf.level}{numeric. Confidence level (default 0.95).}
}
\value{
A data.frame  with eta2, lower and upper confidence limits.
}
\description{
Confidence interval for eta-squared / via noncentral chi-square inversion
}
\examples{
ci_eta2(eta2 = .20, u = 2, D0 = 1000)
}
