% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manymodels}
\alias{manymodels}
\title{Many Models data}
\usage{
data(manymodels)
}
\description{
Simulated data with variables suitable to estimate
several generalized linear models
}
\examples{
data(manymodels)
}
\keyword{datasets}
