% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.lrt}
\alias{power.lrt}
\title{LRT Power}
\usage{
power.lrt(es, prob, df = NULL, N = NULL, sig.level = 0.05, power = NULL)
}
\arguments{
\item{es}{The expected r-square or expected eta-square}

\item{prob}{expected proportions of cases in the dependent variables levels. For dichotomous
dependent variables a number between 0 and 1 (excluded). For categorical variables with more
than two levels, a numeric vector of length equal to the number of levels and whose
values sum to one.}

\item{df}{degrees of freedom of the model or of the effect}

\item{N}{Number of observations. Omit if required sample size N is required}

\item{sig.level}{Significance level (Type I error probability). Set to NULL if required minimal alpha is required}

\item{power}{Power of test (1 minus Type II error probability). Omit if estimated power is required.}
}
\value{
Object of class '"power.htest"' with input and computed parameters
}
\description{
computes power parameters based on the R-squared or eta-squared for
generalized linear model with categorical dependent variables
}
\examples{
r2=.10
p=.5
gzlmpower::power.lrt(es=r2,prob=p,df=2,power=.80)
}
\author{
Marcello Gallucci
}
