% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsizes.R
\name{print.nwa}
\alias{print.nwa}
\title{Print numeric with attributes}
\usage{
\method{print}{nwa}(x, ...)
}
\arguments{
\item{...}{not implemented yet}

\item{object}{object of class "nwa" (numeric with attributes)
or \code{\link[stats]{deviance}} is defined.}
}
\description{
Prints generic named vectors without printing their attributes
}
\examples{
data(manymodels)
model<-glm(ybin~x,family=binomial(),data=manymodels)
print(r2(model))
}
\author{
Marcello Gallucci
}
