% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsizes.R
\name{r2}
\alias{r2}
\alias{r2.default}
\alias{r2.lm}
\alias{r2.glm}
\alias{r2.clm}
\alias{r2.multinom}
\title{Model R-squared}
\usage{
r2(object, ...)

\method{r2}{default}(object, test = FALSE, ...)

\method{r2}{lm}(object, test = FALSE, ...)

\method{r2}{glm}(object, test = FALSE, ...)

\method{r2}{clm}(object, test = FALSE, ...)

\method{r2}{multinom}(object, test = FALSE, ...)
}
\arguments{
\item{object}{object of class "glm", or an object for which the function \code{\link[stats]{deviance}}
or \code{\link[stats]{deviance}} is defined.}

\item{...}{not implemented yet}
}
\value{
an numeric vector with the R-square and adjusted R-square
}
\description{
computes the R-squared for several generalized linear models
}
\examples{
data(manymodels)
model<-glm(ybin~x,family=binomial(),data=manymodels)
r2(model)

}
\author{
Marcello Gallucci
}
