% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{ci_eta2p}
\alias{ci_eta2p}
\title{Confidence interval for partial eta-squared}
\usage{
ci_eta2p(eta2p, u, Dmx, conf.level = 0.95)
}
\arguments{
\item{eta2p}{numeric. the effect size index for predictor x}

\item{u}{integer. Degrees of freedom associated with predictor x.}

\item{Dmx}{numeric. Deviance of the model excluding predictor x.}

\item{conf.level}{numeric. Confidence level (default 0.95).}
}
\value{
A data.frame  with eta2, lower and upper confidence limits.
}
\description{
Confidence interval for partial eta-squared
}
\examples{
ci_eta2p(eta2 = .20, u = 2, Dmx = 1000)
}
