% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsizes.R
\name{eta2_partial}
\alias{eta2_partial}
\title{Partial Eta-squared and Partial Epsilon-squared}
\usage{
eta2_partial(object, ...)
}
\arguments{
\item{object}{object of class "glm", or an object for which the function \code{\link[stats]{drop1}}
or \code{\link[stats]{deviance}} is defined.}

\item{...}{not implemented yet}
}
\value{
an anova table with R
}
\description{
Computes the partial eta-squared and partial epsilon-squared indices for several generalized linear models
}
\examples{
data(manymodels)
model<-glm(ybin~x,family=binomial(),data=manymodels)
eta2_partial(model)
}
\author{
Marcello Gallucci
}
